requirements.txt
Flask==3.0.0
pytest==6.2.4
Werkzeug==3.0.0
httpx
python-dotenv
jinja2
requests